OMAP: mcbsp: Add APIs for raw access to mcbsp registers
The mcbsp library currently supports only a simple single write TX
DMA model that blocks until completion.  In order to implement other
models like audio playback that must keep the DMA filled constantly, we
expose some of the raw registers so that the additional models can be
implemented on top of the existing library.
Also fix error checks of rx_irq which assumed non-zero was a valid
irq number.  Since rx_irq might be set to -ENXIO, the check for a valid
irq should be >= 0 instead.

Change-Id: I2d82be7b7a32dec5d8ffb2f2d8d64ff3ab00e473
Signed-off-by: Mike J. Chen <mjchen@google.com>