Steelhead: Change governor back to interactive.
After last down-integrate from tuna, we changed default governor
to hotplug to match tuna changes.  This has had a detrimental effect
on audio causing many noticeable drops.  Change back to default
to interactive for now.  This might be permanent for us, or we can
try again to match tuna when they improve the behavior of hotplug (this
is a problem TI and system sw is aware of).

Change-Id: I69dbf3aad234f39b575fafc09e3d0c357005eb4e
Signed-off-by: Mike J. Chen <mjchen@google.com>