Wait for DMA to load TXBUF0 during startup.
When starting transmission, wait for the DMA to load the first audio sample
into TXBUF0 before releasing the TX state machine and frame sync generator
from reset.  Failure to do this can cause an underflow condition which seems
to be able to lock up the FS generator so that no data ever gets transmitted.
While this does not seem to happen at 48KHz all that much, it is a pretty
common failure mode when transmitting at 96KHz.  Waiting for the DMA to load
the first sample is something the TRM tells you to do in section 23.8.5.2.1.4
so it seems like a good idea for us to do as well (plus, it make 96KHz
reliable).  Currently, we only wait up to 1mSec total for things to start and
give up with an error if DMA has not loaded the sample by then.

Change-Id: Ide2ddf5bdb4c845aefcc528a08779481586d2d80