Init all bits in the clk divider regs during setup
When setting the 2 divider values to produce the transmit clock, need to make
sure that we explicitly initialize all of the bits in the divider registers.
The existing code was effectively or'ing the divider values with what was
already in the register.  This was causing output sample rate transitions to
fail from time to time.

Change-Id: If5e9bdd3ec1862450cf0cdf2da14b9471b1fe945