Add the ability to register ioctl handlers in ASoC DAIs
Adds the ability to specify an ioctl handler in the DAI ops table for ASoC DAI
drivers.  This allows codec and cpu dais to implement an ioctl handler for
custom extensions like the platform drivers can now.  DAIs who implement these
handlers must be certain to return -ENOIOCTLCMD when they do not recognize and
cannot handle a given ioctl cmd.  When an ioctl reaches the soc_pcm_ioctl, it
will first be sent to the codec_dai, then the cpu_dai and finally the platform
driver (assuming they implement ioctl).  Processing will stop as soon as one of
these components returns a value other than -ENOIOCTLCMD.  If none of these
components handles the ioctl, it will be sent to the default snd_pcm_lib_ioctl.

Change-Id: I9636e1bf06d94742a981b7bb6c4f9ea30a6466b7