ASoC: OMAP PCM - Fix for a bad merge
During initialization of a snd_soc_pcm_runtime, the ASoC core code allocates
substream structures in the snd_pcm member of the runtime based on information
in the dai_link and the codec_dai members of the runtime structure.  If the
dai_link indicates that this runtime is bound to codec DAIs dynamically, then
the decision to allocate playback or capture substreams (or both) is made by the
information in the dai_link (defined by the machine driver).  If the dai_link is
not dynamic, then the decision is made based on the min_channels member of the
playback and capture members of the codec_dai structure (defined by the codec
driver).  In older versions of this code, the codec_dai structure was passed
explicitly to the registered pcm_new member of the platform driver (if any).
The omap-pcm platform driver implementation would then choose to pre-allocate
DMA buffers based on the min_channels field in the playback and capture members
of the DAI structure passed to it.  This could be incorrect if the core code had
made is allocation decisions based on the dai_link information instead of the
codec_dai information.

Liam Girdwood addressed this in the p-android-2.6.35 tree of the omapzoom kernel
repo on 2010-9-15.  On the same day, he also changed the platform driver's
pcm_new op's signature to have a runtime passed to it instead of passing the
codec DAI directly.

Looking at the code here, it seems like the ancestor of this tree merged the
second change (pass runtime instead of codec DAI), but not the first change
(allocate playback/capture PCM structures based on NULL-ness of the substream
structures allocated by the core).  When the second change was merged, in order
to keep checking the codec DAI's playback/capture min_channels member, a pointer
to the codec DAI needed to be fetched from the runtime state.  During the merge,
it looks like the CPU DAI was accidentally fetched from the runtime instead of
the codec DAI.  This led to problems where if the CPU DAI supported both
playback and capture (like the McBSP driver does), but the codec driver does not
(drivers for external amplifiers, for example), the core would end up not
allocating a capture substream (because the codec didn't need one), but the
omap-pcm code would try to pre-allocate buffers for the capture substream
(because the CPU DAI driver did support capture), and it would end up crashing.

Proper fix for this seems to be what Liam originally submitted; pay attn to
whether or not the substream was allocated by the core in order to determine
whether or not to pre-allocate DMA buffers instead of trying to duplicate logic
embedded in the ASoC core.

Change-Id: Ie5884edb202e3dd4a1caaf45c19b39b4d693ae5f