steelhead: Partial bringup of WLAN on Tungsten HW
Changes to support initial bringup of the WLAN portion of the BCM4330 WLAN/BT
module on Tungsten HW.  There are still issues preventing the module from fully
loading which need to be worked out with BCM when they are able to get an
engineer on-site.

Also did a bunch of cleanup/refactoring.  Things like -
+ Removed all traces of the old Panda Wl1271 wireless module.
+ Split bluetooth and WiFi initialization into their own files.
+ Cleaned out the board_mux table since we are now adopting more of a
  set-the-mux-during-function-init style
+ Try to keep GPIO/Resource definitions with init functions instead of leaving
  them at the top of board-steelhead.