Fixes for the OPP table on Steelhead.
Add a config option for disabling L3 clock scaling.  L3 scaling affects some
clocks that are derived from DPLL_PER.  In particular, it changes the rate of
dpll_per_m3x2_ck, which is the source for the auxclk1_ck exported as the
TAS5713 amplifier's MCLK.  The TAS5713 will not function if its clock changes
rate.

Also, make the HSI and AESS OPP entries conditional on the config options for
those units.

Change-Id: I3e9d801491e17f0688ccbcad8474c0fe6e2f9255