ARM: omap4: steelhead: add DDR timing information
Based on the elpida for Panda in pre-evt.  For non-pre-evt, we use
the same settings as tuna for now though that's not tested.  Tuna
created it's own copy but that are the same as what's in
lpddr2_jedec_data.c (which is also the same as in lpddr2_elpida_data.c)
so we don't duplicate but reference those for now.

This is manual integrate of change:

commit 3e5fbbae24a661c6c31e80d389f4f75fc519b03a
Author: Santosh Shilimkar <santosh.shilimkar@ti.com>

    board-tuna: Add Samsung DDR organisation and timings.

    Tuna board memory configuartion is as below

    The memory organisation is as below :
        EMIF1 - CS0 -   4 Gb
        EMIF2 - CS0 -   4 Gb
        --------------------
        TOTAL -         8 Gb

    Change-Id: I3517ab2f52ce0452bbb9fdd011ef6bb2bca56bbd
    Signed-off-by: Ambresh K <ambresh@ti.com>
    Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>

Change-Id: I79627212364c0d7ab5514348925b6249c0ef8a07
Signed-off-by: Mike J. Chen <mjchen@google.com>