ARM: omap4: steelhead: Remove dvi and add gpio_request for HDMI HPD
With the recent merge of tuna-3.0 and new PVR driver, we don't
have the problem of needing a lcd_mgr anymore.

Also, hooked up HDMI HPD gpio because the driver uses it but doesn't
do the gpio_request itself.

With this change, the Android UI is actually coming up on the HDMI port!

Change-Id: I371714ed96329b36420dfcf8d864dfcdccdcd3ab
Signed-off-by: Mike J. Chen <mjchen@google.com>