ARM: omap4: steelhead: Fix HDMI hotplug detection
Earlier tuna merge caused some code to be deleted that we need.  Tuna
doesn't have the level converter chip that Panda and Tungsten have.
Panda code also wasn't quite right in that it would disable the HPD
converter on the level converter on disconnect, but then we'd never
detect the next insertion.  Now always enable the HPD converter on
boot, and only disable/enable the level converter on HDMI detection.

Change-Id: Ifc1936e81e4f2cb47836ebb880213f4470745522
Signed-off-by: Mike J. Chen <mjchen@google.com>