pn544: Fully qualify the gpio_request.
The earlier gpio_request() usage didn't specify gpio direction or initial
value.  On Crespo, that was okay because Samsung c100 processor gpio
library and board file for herring left it configured after init.

On omap, gpio_free() always resets the status to unconfigured so there's
no way for board file to configure it and still allow the driver to do
a simple gpio_request().  The driver MUST fully specify direction and
level.

Change-Id: I54c6a99aba2f714d10d053198824017a9000548a
Signed-off-by: Mike J. Chen <mjchen@google.com>