Don't program reserved (must be 0) bits to 1s.
Found that some of the bits which the TAS5713 spec had flagged as
reserved-must-be-0 in the volume slew register were getting programmed as 1s
instead.  The chip didn't seem to care, and there was no noticable change in
behavior, but I felt that its always better to match the spec whenever you can.

Change-Id: I5ad1d3cac76af7d86b429c389b8fb5c8ab6b50c4