Fix the TAS5713 reference clock.
On steelhead builds, change the allowed frequencies for the output of the
DPLL_PER M3 dividers for both of the defined OPPs to be 61.44MHz instead of
allowing them to switch between 192 and 256MHz.  This is the divider which feeds
our aux clk 1 fref output which drives the TAS5713's mclk signal and must remain
stable at all times.

Change-Id: I5d6d8c2a8b89cc1cf47484ed811fd51d5316a2b7