ARM: omap4: steelhead: Fix hw_rev identification and use the info for pinmux
hardware_rev identification was broken because the input lines needed
pull downs (the boards just use stuff/no-stuff resisters to 1.8V, and when
not stuffed, the lines would be floating).

Renamed the pre-evt to alpha to be consistent with what HW team is doing.

Use hardware_rev to make pin muxing and other setup decisions that we'd
earlier used cpu_is_omap446x() for because we could have omap44xx SOCs
in some EVT boards if power issues are unsolved.

Change-Id: I40cfdf00283ac63e7373292a41f41ff5c230cc51
Signed-off-by: Mike J. Chen <mjchen@google.com>