Revert "OMAP4: HWMOD: Update thermal alert irq to 127"
This reverts commit 6fed58ca27550c7d02cce0b6b6df42155fd35bc5.

Revert for now because we don't have the PPA that does the routing
of the thermal interrupt to irq 127 and we're not getting the irq 126
otherwise.  Also, GP has no PPA so could not do this routing.  We
can conditionalize on GP vs PPA, but until we have the new PPA, it's
easier just to revert this.

Change-Id: Ieb8c68e06f3c643b32a9b6c51ac644bcfde93917
Signed-off-by: Mike J. Chen <mjchen@google.com>