steelhead: Turn on IGMP (and other mcast stuff)
CONFIG_IP_MULTICAST was not being turned on in the steelhead kernel.
Amongst other things, this was causing IGMP membership reports to not
get sent.  This was causing issues witht he A@H lab network which
apparantly is IGMP aware and will not forward wired traffic to the
wireless network unless there is at least one wireless client
participating in IGMP and interested in the traffic.

Exec. Summary: If you are going to do multicast in a modern network,
you really need to be doing IGMP.

Change-Id: Icc723043efe49c32a402ea37253cbef91f39a384
Signed-off-by: John Grossman <johngro@google.com>