OMAP:DSS/DSSCOMP: Remove race conditions on callback setting and error handling
This fixes lockups and crashes observed during display error
injection tests.

Wrapping prior callback for managers created race condition,
where the manager settings could change between setting
the manager info and wrapping the callback.  Since callbacks
are never set on the original managers, this wrapping functionality
is removed.

The error handling pathway was also prone to race conditions as
the settings could have gotten applied between the mgr->apply failure
and the unsetting of the callback.  Therefore, created a protected
method to unregister a callback (given that it has not been applied).

Finally, callbacks were not called when set_mgr_info failed, but
mgr_apply succeeded.  This is now needed as we are not wrapping the
callback outside of set_mgr_info.

Also fixed two bugs in mgr_blank() where the apply ISR was not
registered, and where we did not wait got GO if GO was not already
present during configure().

Change-Id: Ifb4e07ea3bc17006c394517b8eb216ebfe154b73