WAR:ARM:omap4:steelhead: Ensure we have enough bus bandwidth for EHCI
We were seeing trouble with EHCI transactions and traced the cause to
the controller being unable to fetch the payload for their transaction
when L3 was running at OPP50.  When the controller was unable to get
data from L3, it would abort its bulk OUT transaction, corrupt the CRC
it was sending to the USB device, and try again next time (this is all
consistent with the EHCI specification).  Trouble was, it would almost
certainly fail the next time through as well because L3 was apparantly
running too slow.  In the worst case, this would basically wedge the
entire outbound EHCI endpoint (although it would not affect other
endpoints).

Adding a bus throughput constraint ensures there is sufficient
bandwidth available any time we are using the EHCI controller.

Change-Id: I128cb72d29820f8d275123cdf2e2a85b26c9c169
Signed-off-by: John Grossman <johngro@google.com>