TAS5713: fix the output mux to match tungsten.
The way the audio board for tungsten is layed out does not match the
way we were configuring the output mux for the TAS5713.  To be
specific, the right hand channel (OUT_C and OUT_D) needed to have its
polarity flipped.  Now that everything is properly in phase, you ought
to be able to smash a window with a max volume sine wave at the proper
frequency.

Change-Id: I400511269db670ead72b2025c2e51744f4369a2b
Signed-off-by: John Grossman <johngro@google.com>