usb: ehci: omap: increase IN/OUT packet buffer thresholds
This is a variant of change ID
I3625f7034053ed5602a559a0542b9561a9e9a1ec originally pushed to
https://partner.source.android.com/kernel/private/omap-samsung on the
android-omap-tuna-3.0 branch by Vikram Pandita <vikram.pandita@ti.com>

This change increases the EHCI transfer thresholds to 512 bytes, the
maximum size of a bulk data transaction for USB 2.0.  This means that
the EHCI controller will never start a bulk data send until it has all
of the data buffered in the EHCI controller.  Likewise, it will not
try to push data to system memory from RX until it has received the
whole transaction, although this is less important overall.

Now, no matter how badly things go with the L3 feeding data to EHCI,
we will never get into a situation where the EHCI controller
underflows during an OUT transaction and has to cancel and retry the
transaction.  This has a negative effect on the theoretical throughput
of the system, but even conservative estimates of the new hypothetical
throughput (~195Mbps) indicate that they are more than enough for
Steelhead/Tungsten (which dedicates this bus to a single 100Mbps
Ethernet controller).

Change-Id: Idf38a4a45e9f3fe88ec1a38533a602a45a282104
Signed-off-by: John Grossman <johngro@google.com>