steelhead_avr: Pruning and adding new features
Change some pr_info() to pr_debug() in the key reading code
since it's too noisy now if that the encoder works.

Also ignore i2c errors for now in the key reading.  It appears
to happen when rotation a lot.  Theory is that the AVR is too
busy handling encoder interrupts to keep up with i2c.  Ignoring
seems harmless.

Remove bank API that we don't think is useful.

Add mute led support.

Add ioctl to reset the AVR via a board file provided GPIO.

Also support a reset key event notification that the AVR
will send on boot so we can tell if the AVR has rebooted
itself (or was requested to reboot during fw update).

Also to match UI spec, kernel shuts off mute led on probe
or reset detection, but doesn't stop boot animation.  It lets
userspace do that.

Also, AVR will be changed so interrupts and button events
will be enabled on boot and no commands are needed by the
kernel to enable those.

Don't cache the firmware/hardware type/rev info since that can
change after a firmware update.

Don't fail if we can't talk to AVR in probe (when setting mute led).
We still need the dev node to allow the user space avr_updater to
reset the AVR into it's bootloader for UART flash updating.

Change-Id: Ifb9a20b13f72e0bc76ab25e5dc4e61adc6dca51d
Signed-off-by: Mike J. Chen <mjchen@google.com>