ARM: omap4: steelhead: Add flag value for a normal reboot request
In order to implement the special case of losing power while in
recovery and wanting the next boot to go into recovery again,
I need to be able to distinguish a request for a normal reboot
vs a cold boot (not request).  A normal reboot is what recovery
asks for after it is done.

Change-Id: I0f65332a09cc0b1fe2d33a297dda0d75509122e4
Signed-off-by: Mike J. Chen <mjchen@google.com>