Steelhead: ASoC: Instantiate a TAS5713 audio "card"
Add the ASoC machine driver for steelhead to bring together the
various pieces which make up the audio pipeline which eventually leads
to the TAS5713 audio amplifier.  Right now the TAS pipeline will start
up and run, the McBSP will be producing a valid I2S signal, but no
audio will be produced as the TAS codec driver is currently just a
stub.

Note: instead of adding the TAS5713 as another audio device on the existing
card, the pipeline was models as a new card with a single device.  This is
because ASoC will not instantiate a card if any of its devices have trouble
instantiating.  We don't one trouble with one of the audio paths preventing the
other from loading, so we set them up as different cards instead of different
devices on a single card.

Change-Id: Iabbf3c6e569e6df69b8f3b632aab06924fdc7989
Signed-off-by: John Grossman <johngro@google.com>