Steelhead: Take McBSP1 output pins out of safe mode.
McBSP1 is used by tungsten to drive I2S into the BCM4430.  Usually,
this is used for BT headset profile.  Right now, tungsten does not
implement this as a feature.  Instead of leaving these pins in safe
mode, we are supposed to be setting them to outputs which are driven
low.  For now, we achieve this by setting these pins up as GPIO
outputs being driven low.

Change-Id: Ia8598d17f917638de25dd198f27eb8bf73285ce0
Signed-off-by: John Grossman <johngro@google.com>