ARM: omap4: opp: For steelhead, set l3 opp50 voltage to nominal
Steelhead has a usbhost controller requires a higher voltage
to meet timing requirements than opp50 currently defines.
We use that usbhost controller to communicate with an always
on smsc ethernet chip.

Instead of holding a bandwidth constraint to force opp100 (and it's
higher voltage), we change the opp50 voltage to nominal so at least
we can drop the clock rates for dplls in the domain.

Change-Id: I535e68541696da3cf3765a745b2bb82159a74c33
Signed-off-by: Mike J. Chen <mjchen@google.com>