OMAP: HDMI: Attempt to reset DDC I2C in the case of timeout.
On a small number of occasions, I have managed to get my DDC I2C
controller into what appears to be a stuck state.  The clock and data
lines were being held low by OMAP, the DDC_STATUS register indicated
that the FIFO was full and that there was an operation in progress,
and the DDC_CMD register showed that the last issued command was 0xA
(Clock SCL).  The condition seemed sticky.  Every time the driver
would attempt to reset the I2C bus using a general reset (the Clock
SCL command), the in-progress bit in DDC_STATUS never cleared, so the
driver would time out and fail the read.  After 5 failed attempts to
read the EDID, the system would just give up.

Issuing an abort transaction command (0xF) to the DDC I2C controller
while it was in this state seemed to clear the condition.  The SCL and
SDA lines were immediately released and the next attempt to read EDID
succeeded.

The condition is extremely difficult to reproduce.  I have only seen
two confirmed instances of it in the last 5 weeks.  There does not
seem to be any downside to attempting to abort a transaction in
progress if commands like Clock SCL are timing out, and it clears the
stuck controller condition (at least in my case), so go ahead and
attempt an abort-transaction command in the case of a timeout.

Change-Id: I57e4c8a4eb15f1b335bb68a0ca2abe1f5984c208
Signed-off-by: John Grossman <johngro@google.com>