Steelhead: TAS5713: Fix async power sequencing.
Now that power sequencing for the TAS5713 is asynchronous, we need to
add some hooks to allow board specific code to run at specific points
in the power sequencing operations.  In specific, on steelhead, it is
important that the external level translator become enabled before
power up and remain enabled until after shutdown.  The way the code
was before, this was being handled by the machine driver in response
to ASoC callbacks which were causing the level translator to turn on
and off in the middle of asynchronously executed startup and shutdown
sequences.  This chage moves control of the LT interface enable out of
the board specific machine driver, and into callback functions passed
to the TAS5713 driver via pdata allowing the TAS5713 driver to notify
the board specific platform code just before power up and just after
power down of the amp.

Change-Id: I013cc4c44c33f6850950540e5e264e12d3035827
Signed-off-by: John Grossman <johngro@google.com>