Update comment to reflect the results of TI's investigations.
There was some disagreement between what the TRM said about accessing
the 32k count register and what TI's code was doing.  I asked TI to
look into it, and they confirmed that the docs are wrong and the code
is correct.  This change updates the comment around this access to
document that TI looked into this and said everything was OK so that a
future engineer with old documentation won't be confused.

Change-Id: Icf7fbbbbec21b00b966235627a27b405e852735a
Signed-off-by: John Grossman <johngro@google.com>