OMAP: DSSComp: Add a forgotten break statement.
The IOCTL switch for dsscomp was missing a break statement in the
SETUP_DISPLAY handler causing it to fall through to the default
handler which was always reporting an error of -EINVAL (even when
SETUP_DISPLAY succeeded).  Add the break in so that folks who call
SETUP_DISPLAY and check their error code don't end up confused.

Change-Id: I94fa2d6c84fcd211fde948b5cd60bc837365793b
Signed-off-by: John Grossman <johngro@google.com>