Steelhead: HDMI: increase max allowed pixel clock.
Increase the maximum allowable HDMI pixel clock from the 75MHz copied
from Tuna to 165MHz.  165MHz is the max allowed by HDMI 1.3 and is the
max the OMAP HDMI unit is capable of.  Tuna needed to limit this to
75MHz because they needed to serialize their HDMI over an MHL
interconnect (which only supports up to 75000).  Tungsten has no such
restrictions.  Bumping the max pixel clock up allows us to output a
full 1080p60 signal (which is a good thing when we are trying to pass
HDMI compliance testing.)

Change-Id: I3a160f17f54e095ff712ce10b4f0aff30ec30589
Signed-off-by: John Grossman <johngro@google.com>