steelhead_avr: Add i2c retries
The AVR has a simple i2c controller and runs a single threaded
loop with asynchronous interrupts.  The upshot is that it can
sometimes be unable to respond to i2c requests for brief
periods of time (e.g. when handling one interrupt that holds off
the i2c handling interrupt) so we add some retries to the i2c
transfers.  I don't see any more failure messages with a retry
count of five.

Change-Id: Ib7662af4fc5c890ad18d108a27d01a77add74b9a
Signed-off-by: Mike J. Chen <mjchen@google.com>