ARM: omap4: steelhead: Set source of gpu_fck to dpll_per
Partial fix for bug b/5800242.  The source clock specified in
the OPP for gpu_fck is DPLL_PER, but source config register
wasn't setup to make it so (default source if DPLL_CORE).

Not sure how tuna did it, maybe in the bootloader, but this
patch sets the source in the board file.  TI might give a
better fix later.

Change-Id: I9d5e00e552d1c906598f02bac9d1e7b9ad8d6409
Signed-off-by: Mike J. Chen <mjchen@google.com>