ARM: omap4: Constrain DPLL_CORE to always operate at OPP100 voltage
This is a merge of two patches received from TI via email.  Official
changes will probably come in from upstream branches later but
I wanted to get things in for testing sooner because it can impact
thermals.

The two parts were:
a) lock the core voltage domain - I believe to the OPP100 level because
that's the initial OPP.
b) adjust the smart reflex setting for OPP50 case to be the same
as the OPP100 case to make sure smart reflex doesn't decide on
the wrong operating voltage for the OPP50 case.

Partly fixes bug b/5800242.

Change-Id: I16160896bdd6bafcf594518fd23e561f92cfc14b
Signed-off-by: Mike J. Chen <mjchen@google.com>