Steelhead: HDMI: Always power the DDC/CEC LCs
For now, always power the level converters for the DDC and CEC busses.
Current, the HDMI compliance testers feel that this is a requirement,
even though there is no obvious reason why this is needed and it just
wastes power.  For now (while our standards folks lobby to get the
spec changed), we just leave it powered at all times.

Change-Id: I3bed99b7ff7db813c64d0caecc4940380c48a395
Signed-off-by: John Grossman <johngro@google.com>