OMAP4: clock: Disable drift guard.
Drift guard bit in all DPLLs should be disabled in production
software, as it is indented for hw development. With this bit
disabled CPU would be informed of reason for losing DPLL lock
for any the reason like electrical, noise, temperature out of
range.
If this bit is set, then DPLL's will be relocked after loosing
the lock for any of the mentioned reason. And in certain cases
some peripheral devices might malfunction, if CPU is not
informed this relock.

Disabling this bit has no logical impact and usually disabled
for all DPLLs in boot-loader. But its better to have it checked
& disabled explicitly in kernel.

Change-Id: I56c4d60610f5ac0bbf90924a372e42bcbed787ec
Signed-off-by: Ambresh K <ambresh@ti.com>
Signed-off-by: Vinay Chaurasia@ti.com>