ARM: omap4: steelhead: Increase non-secure tiler heap size
We'd been observing PA secure violations when playing DRM videos.
Investigations showed that we were failing to allocate the framebuffers
from the non-secure tiler heap.  The reason is that although we're
using the same panel dimensions as tuna, our orientation is reversed.
This results in larger memory requirements because of alignnment code
in omaplfb_displayclass.c that rounds the x_res to multiples of 0x400 (1024).

For tuna, with x_res=720 (& y_res 1280), data->w = ALIGN(720,1024) * 3 = 3072.
This is processed by omap_tiler_alloc to generate n_phy_pages to be 3840,
which is 15MB, the exact size of the non-secure tiler heap tuna allocated.

For tungsten, using the generic_720p panel, with x_res=1280 (& y_res=720),
data->w = ALIGN(1280,1024) * 3 = 6144.  omap_tiler_alloc() computes
n_phy_pages to be 4416, which is a little more than 17MB.  We round this
up to 18MB.

To keep the memory size the same, we could swap the dimensions in
generic_720p panel, so it's the same as Tuna, but that causes our
boot animation to get rotated 90 degrees.  We could try to optimize
this by rotating the boot animation, but it really would be lie/hack.

This bug was hidden because OMAPLFBInitFBDev() falls back to allocating
from the secure tiler heap when the non-secure alloc fails.  But using
a secure heap ptr trips the SMC PA when playing a DRM file, which caused
the errors we were seeing.  The only symptom was a kernel log that I
noticed when TI told me where the ptr was that was causing the SMC PA
to complain:

  omap_tiler_alloc: failed to allocate pages to back tiler address space

Change-Id: Idff34c6dfc7cf321cc4047baa6f066cd1455674a
Signed-off-by: Mike J. Chen <mjchen@google.com>