ARM: omap4: steelhead: Remove two sample modules from build
I noticed we have some modules still being built (around 4 of them)
so I investigated what they were.  Two of them appeared to be
sample code, so I'm disabling them in this change.

Two others are a bit more unclear.  One is scsi_wait_scan, which
doesn't seem disable-able (at least setting it to "is not set"
in steelhead_defconfig did nothing to change it from being built
as a module).

The other, ansi_cprng, I'm not sure if we use or not so I decided
not to touch for now.

Change-Id: Ibd6ca77e76dcab15aec2e1f7c031b55e3c1937d7
Signed-off-by: Mike J. Chen <mjchen@google.com>