HDMI: introduce a new switch to signal support for audio.
Introduce a new sysfs switch ("hdmi_audio", in addition to the
original "hdmi") to signal support for HDMI audio to the userland
software.  This makes it a little easier for the higher level software
to properly sequence the bring-up of audio on an HDMI connection.
Strictly speaking, until a video mode is chosen, you cannot be certain
that an HDMI connection can support a selected audio mode (even one
advertised in the EDID information).  Before this change, the HW
composer HAL and the WiredAccessoryObserver were in a race, the HW
composer trying to set a video mode in response to a hot-plug event,
and the WiredAccessoryObserver trying to signal support for HDMI audio
to the audio HAL.  If the audio HAL ended up trying to bring up the
ALSA HDMI audio driver before a video mode had been selected, the
driver would refuse to load.

Also, add support for parsing audio modes from the CEA extension
blocks of an HDMI sink's EDID blocks.  Just because an HDMI sink has
been connected does not mean that the sink supports any audio modes.
Before this, the code was not even bothering to parse the supported
audio modes.

Now, the WiredAccessoryObserver can pay attention to the "hdmi_audio"
switch instead of the "hdmi" switch.  The "hdmi_audio" switch will be
triggered only after a video mode has been successfully configured,
and only when the sink's EDID claims support at least some audio mode.

There is still a lot of room for improvement here.  Right now, the
HDMI audio driver is still just checking for the presence of an HDMI
sink and not bothering to take a look at the modes supported by the
sink.  This should be OK so long as the HDMI audio driver continues to
only support the basic PCM audio modes, but to properly support
multichannel PCM and compressed passthru audio formats, the EDID needs
to be checked and the selected audio mode needs to be checked against
the selected video mode's pixel clock.

see bug 6023647

Change-Id: I37e5ed3e4de4d0b84407ac0846a8b10c8dcce9e6
Signed-off-by: John Grossman <johngro@google.com>