HDMI: clean up the HPD state machine some.
The issue here is that if a series of hot plug events occurred while
the HPD worker was executing a long running operation (like reading
EDID), and the final state after the series of events was the same as
at the start, the HPD state machine would only partially reset.  One
side effect of this was that higher levels of code were not being
properly informed that the line was un-plugged and re-plugged again.
Hypothetically, if an unplug/replug sequence happened during an EDID
read and was the result of an AVR switch changing inputs, the system
could end up improperly using EDID information from an old switch
position on a new switch position.

This change avoids that possibility by changing the state machine so
that the HPD IRQ handler always forces the machine into the "RESET"
state instead of putting the machine into the "OFF" or "START" state
depending on the state of the HPD line.  The HPD worker is now
responsible for checking the state of the line and attempting to bring
the connection up if HPD is asserted.  If a new HPD event shows up
while the worker is in the process of doing work, the worker will
always be forced back into the reset state and always make certain
that the "hdmi" switch is properly set to unplugged before proceeding
to eventually read EDID and set the switch to plugged if everything
looks good.

see bug 6023647

Change-Id: I3d125c6b20ecad5cc211a75c4f3e03467f28da7a
Signed-off-by: John Grossman <johngro@google.com>