OMAP4: OMAPFB : Add support for default display scaling.
Currently it is assumed that frame buffer size
is same as default panel size. With this assumption
GFX pipeline is always presented to default display.
GFX pipeline cannot scale so we cannot have a frame
buffer which can be larger and cloned on to HDMI
and scaled down on default display.

Also in cases like default HDMI where panel resolution
will depend on TV we cannot assume GFX pipeline
to connect to framebuffer.

Change-Id: I9e2d40bed34c5fbf2bac31668d210e0bc586bb40
Signed-off-by: Dandawate Saket <dsaket@ti.com>
Signed-off-by: Muralidhar Dixit <murali.dixit@ti.com>