Fix bug in musb when booting as host.
The musb driver had a bug when the kernel boots while devices
are attached to the bus. omap2430_musb_enable does not set musb in
host mode, even when it detects the ID pin.

The fix calls musb_platform_set_vbus in musb_start when the phy
is in A mode. This effectively puts musb in host mode, even if
the platform init does not.

Change-Id: I7abe0cf45d1beaf5dc1d92bf82840e5f2873cfd5