OMAP4: McASP: HACK: Force 200MHz L3 bus clock.
As it happens, the McASP has a much smaller buffer than the McBSP does
for audio samples.  The McBSP has 128 samples, or 64 time slots in
stereo mode, while the McASP has only 1 sample.  This translates into
an 11uSec latency requirement for McASP when running with 48KHz stereo
audio, vs 1.33mSec for the McBSP.  Unfortunatly, when running the L3 at
100MHz, the system seems unable to keep up sometimes when graphics are
being clocked out for 720p, and things get worse when we go up to
1080p.

If the McASP does not get its data from sDMA in time, it can lock up
causing the DMA to stall.  This causes ALSA to stall for 10 second,
which blocks the entire audio HAL and mixer for 10 seconds.

Locking the L3 at 200MHz seems to be the best we can do to avoid the
issue for now.  This change adds a hack throughput constraint to force
the L3 to run at at least 200MHz at all times.

Change-Id: I004e45746e99a64c048b2cca5ea030e69f8ae9b4
Signed-off-by: John Grossman <johngro@google.com>