OMAP: HDMI: mark preferred mode before passing to user space
Define two new flags for the modedb records that
get passed into user space. Mark the EDID preferred
mode and modes that are actually possible according
to the driver.

This allows user space to see all of the modes that the
panel said it could do, even if not possible. A better-
informed user space can determine which mode to go to
a bit better.

Change-Id: Ifb69dbbe7097ba003ed5228c8d4a492b6128eb92