OMAP4: HDMI: Fix race condition for DDC
hdcp and hdmi both use DDC (the former for hdcp authentication
and the latter for edid reading) and they both have state
machines that right now can operate in parallel.  This can
cause problems, such as a bad edid read which can cause hdmi
audio to fail.

This change is based on a patch from TI, to start the
hdcp state machine after the hwc in userspace has called the
setmode API (sets the custom_set flag), which is after edid
has already been read.  This should effectively serializes hdcp to
be after edid reading so both don't happen in parallel.

It's may be possible if the cable is unplugged and replugged before
the hdcp state machine (which operates as a series of delayed work
events) completes so that it's still running when the next edid
read starts, but should in general fix the issue for the typical
case.  We're likely to have other unplug/plug issues that will
have to be flushed out with more testing.

Change-Id: Ib97e292494a9731adb7444c114abd7427e7e9834
Signed-off-by: Mike J. Chen <mjchen@google.com>
(cherry picked from commit 77f18c8faa6545a7e29ca04930a7415f51cab068)