OMAP4: HDMI : Fix issue with context save restore for HDMI default display
When HDMI is default display clocks for dispc and dss are only requested
by HDMI driver. On a hot plug out the clocks are release immediately and
since its the only display reference count will also be zero. So we will
save the context with video pipelines active and pushing content to hdmi
On restore the base address of video pipelines will not be valid and we
will get a L3 error for DMA access since attribute register in pipeline
is set.

Solution is to blank out the hdmi before context save so that we will
have a clean state of overlays when context is saved.

Change-Id: I28f0f132883ffa48c3f4ff52f922cbf0de5ed081
Signed-off-by: Dandawate Saket <dsaket@ti.com>
(cherry picked from commit c4d249a2a3df547215c22ec340624631a706e73c)