HACK: bluetooth: steelhead: Set default idle_timeout to 500ms
This engages sniff mode after 500ms of idle.  The variable
is also available to set via sysfs, but it's created late
in boot when userspace bluez creates the hci connection
so the permissions are root:root and can't easily be set to
anything else so this is the easiest way to get us to sniff
mode quickly during setup, which will help in particular with
setup (makes us drop less multicast beacons).

Change-Id: Iee2ab89f0f8f5fe2612691953058d69a82339dec
Signed-off-by: Mike J. Chen <mjchen@google.com>
(cherry picked from commit 409164fbd737f29daaca435cb54326910058245b)